+++ backend/src/main/java/com/luiswiederhold/backend/exception/LowConfidenceException.java
@@ -1,4 +1,4 @@
-package com.luiswiederhold.backend.flashcards.exception;
+package com.luiswiederhold.backend.exception;
 
 public class LowConfidenceException extends Exception{
     public LowConfidenceException(String exception) {
