+++ backend/src/main/java/com/luiswiederhold/backend/flashcards/FlashCardController.java
@@ -2,12 +2,11 @@ package com.luiswiederhold.backend.flashcards;
 
 import com.luiswiederhold.backend.flashcards.DTOs.FlashcardContentDTO;
 import com.luiswiederhold.backend.flashcards.DTOs.FlashcardSideDTO;
-import com.luiswiederhold.backend.flashcards.exception.LowConfidenceException;
+import com.luiswiederhold.backend.exception.LowConfidenceException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
